+++ apps/next-app/CHANGELOG.md
@@ -1,5 +1,26 @@
 # next-app
 
+## 0.1.67
+
+### Patch Changes
+
+- Updated dependencies [7e3716c]
+  - @nation-a/ui@0.16.21
+
+## 0.1.66
+
+### Patch Changes
+
+- Updated dependencies [96d56b5]
+  - @nation-a/ui@0.16.20
+
+## 0.1.65
+
+### Patch Changes
+
+- Updated dependencies [886ce9b]
+  - @nation-a/ui@0.16.19
+
 ## 0.1.64
 
 ### Patch Changes