+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nation-a/tokens
 
+## 0.4.1
+
+### Patch Changes
+
+- 886ce9b: Added background v3
+
 ## 0.4.0
 
 ### Minor Changes